Drop old walk module.
